+++ FreshMinds/src/main/java/de/freshminds/manager/TransactionManager.java
@@ -9,7 +9,7 @@ import de.freshminds.main.Core;
 
 public class TransactionManager {
 	
-	public void create(int transactionNumber, String username, int articleNumber, int amount, int price, String paymentMethod, Date timestamp) {
+	public void create(int transactionNumber, String username, int articleNumber, int amount, double price, String paymentMethod, Date timestamp) {
 
 		Transaction transaction = new Transaction(transactionNumber, username, articleNumber, amount, price, paymentMethod, timestamp);
